requirements.txt
pyserial==3.5
python-dotenv==1.1.0
tk==0.1.0
